completed scheduler

+++ models/major.py
@@ -5,4 +5,4 @@ from typing import List, Set
 class Major:
     name: str
     major_courses: Set[str]        # course codes
-    credit_required: int  
+    credit_required: float
